implements asset bulk import
