+++ src/Items/ItemDetail.js
@@ -19,6 +19,7 @@ var Modal = Bootstrap.Modal;
 var Button = Bootstrap.Button;
 var Form = Bootstrap.Form;
 var moment = require('moment');
+var fileDownload = require('react-file-download');
 
 import {restRequest, checkAuthAndAdmin} from "../Utilities.js";
 import CartQuantityChooser from '../ShoppingCart/CartQuantityChooser';
@@ -55,6 +56,7 @@ class ItemDetail extends React.Component {
     this.logItemQuantityChange = this.logItemQuantityChange.bind(this);
     this.onRowClickCart = this.onRowClickCart.bind(this);
     this.clearAlert = this.clearAlert.bind(this);
+    this.exportAssets = this.exportAssets.bind(this);
   }
 
   getDetailedItem(id, cb) {
@@ -162,6 +164,16 @@ static editGetResponse(data) {
   return data;
 }
 
+exportAssets() {
+  restRequest("GET","/api/item/asset/csv/export?item_id=" + this.state.itemData.id, "application/json", null,
+              (responseText)=>{
+                console.log("Successfully got assets");
+                fileDownload(responseText, this.state.itemData.name + 'Assets.csv');
+              }, (status, errResponse)=>{
+                console.log('Failed to get assets');
+              });
+}
+
 openModal() {
   this.setState({showModal: true}, ()=>{});
 }
@@ -266,10 +278,6 @@ renderDisplayFields() {
     );
   }
 
-  testOnHide() {
-    console.log("TEST ON HIDE");
-  }
-
   render() {
     if(this.state.itemData == null) return null;
 
@@ -303,7 +311,12 @@ renderDisplayFields() {
         null
         :
         <div>
-        {(isStaff && this.state.id != null && this.state.isAsset) ? <AssetTable id={this.state.id} updateCallback={this}/> : null}
+        {(isStaff && this.state.id != null && this.state.isAsset) ?
+          <div>
+          <AssetTable id={this.state.id} updateCallback={this}/>
+          <Button onClick={this.exportAssets} bsStyle="primary">Export Assets</Button>
+          </div>
+          : null}
         <br />
         <p><b>Outstanding disbursements containing this item: </b></p>
         {this.generateItemStackTable(this.state.itemData.outstanding_disbursements)}
